requirements.txt
elasticsearch
ollama
rich
atlassian-python-api
pandas
markdown